basic responsive primary layout
