+++ src/components/icons/DownTriangleIcon.vue
@@ -1,5 +1,5 @@
 <template>
-  <Icon fill="#313C59" class='icon'>
+  <Icon :fill="white ? '#FFF' : '#313C59'" class='icon'>
     <path d="M.07 0C0 0-.02.03.02.07l4.9 4.9c.05.04.11.04.15 0L9.98.07C10.02.03 10 0 9.93 0H.07z" />
   </Icon>
 </template>
@@ -12,7 +12,10 @@ export default {
   name: 'DownTriangleIcon',
   components: {
     Icon
-  }
+  },
+  props: {
+    white: Boolean
+  },
 }
 </script>
